feature/note-2-add-user
refactoring database connect
