+++ src/config/typeorm.config.ts
@@ -1,25 +1,23 @@
-import { getDatabaseOptions } from '@app/config/database.constants';
-import {
-  DatabaseEnum,
-  DatabaseScopeInterface,
-} from '@app/types';
 import * as process from 'process';
 import path from 'path';
 import dotenv from 'dotenv';
 import { DataSource } from 'typeorm';
+import { SnakeNamingStrategy } from 'typeorm-naming-strategies';
 
 dotenv.config({ path: path.resolve('./.env') });
 
-const config: DatabaseScopeInterface = {
-  type: process.env.DB_TYPE as DatabaseEnum,
+export default new DataSource({
+  type: 'postgres',
   host: process.env.DB_HOST.toString(),
   port: parseInt(process.env.DB_PORT, 10),
   username: process.env.DB_USERNAME.toString(),
   password: process.env.DB_PASSWORD.toString(),
   database: process.env.DB_NAME.toString(),
   logging: !!(parseInt(process.env.DB_LOGGING, 10)),
-};
-
-const options: DatabaseScopeInterface = getDatabaseOptions(config) as DatabaseScopeInterface;
-
-export default new DataSource(options);
+  synchronize: !!(parseInt(process.env.DB_SYNCHRONIZE, 10)),
+  entities: [path.resolve(__dirname + '/../modules/*/*.entity{.ts,.js}')],
+  migrations: [path.resolve(__dirname + '/../migrations/*{.ts,.js}')],
+  migrationsTableName: '__note_migrations',
+  migrationsRun: false,
+  namingStrategy: new SnakeNamingStrategy(),
+});
