--- src/types/enums/database/database.enum.ts
@@ -1,4 +1,0 @@
-export enum DatabaseEnum {
-  MYSQL = 'mysql',
-  POSTGRES = 'postgres',
-}
