+++ src/types/index.ts
@@ -1,2 +1,1 @@
-export * from './enums';
 export * from './interfaces';
