--- src/types/interfaces/database/database-scope.interface.ts
@@ -1,13 +1,0 @@
-import { DatabaseEnum } from '@app/types';
-
-export interface DatabaseScopeInterface {
-  type: DatabaseEnum;
-  host: string;
-  port: number;
-  username: string;
-  password: string;
-  database: string;
-  synchronize?: boolean;
-  logging?: boolean;
-  debug?: boolean;
-}
