--- src/types/interfaces/database/index.ts
@@ -1,1 +1,0 @@
-export * from './database-scope.interface';
