+++ src/types/interfaces/index.ts
@@ -1,2 +1,1 @@
-export * from './database';
 export * from './user';
